-count returns NSUInteger which can be int32_t or int64_t depending on architecture and requires different token in -stringWithFormat (https://developer.apple.com/library/ios/documentation/Cocoa/Reference/Foundation/Miscellaneous/Foundation_DataTypes/index.html), use NSInteger instead of int

+++ TFMMobileMarketManager/ImportTool/ImportTool.h
@@ -39,9 +39,9 @@ typedef NS_ENUM(NSInteger, ImportDumpOptions)
 
 - (id)initWithSkipSetting:(bool)skipFirstRow;
 
-- (unsigned int)importStaffFromCSV:(NSURL *)url;
-- (unsigned int)importVendorsFromCSV:(NSURL *)url;
-- (unsigned int)importLocationsFromCSV:(NSURL *)url;
+- (NSUInteger)importStaffFromCSV:(NSURL *)url;
+- (NSUInteger)importVendorsFromCSV:(NSURL *)url;
+- (NSUInteger)importLocationsFromCSV:(NSURL *)url;
 
 - (bool)importDump:(NSURL *)url;
 // - (bool)importTable:(NSURL *)url;
